+++ woodwork/logical_types.py
@@ -4,7 +4,7 @@ import warnings
 import pandas as pd
 import pandas.api.types as pdtypes
 
-from woodwork.accessor_utils import _is_dask_series, _is_koalas_series
+from woodwork.accessor_utils import _is_dask_series, _is_spark_series
 from woodwork.config import config
 from woodwork.exceptions import (
     TypeConversionError,
@@ -14,13 +14,14 @@ from woodwork.exceptions import (
 from woodwork.type_sys.utils import _get_specified_ltype_params
 from woodwork.utils import (
     _infer_datetime_format,
+    _is_valid_latlong_series,
     _reformat_to_latlong,
     camel_to_snake,
     import_or_none,
 )
 
 dd = import_or_none("dask.dataframe")
-ks = import_or_none("databricks.koalas")
+ps = import_or_none("pyspark.pandas")
 
 
 class ClassNameDescriptor(object):
@@ -54,7 +55,7 @@ class LogicalType(object, metaclass=LogicalTypeMetaClass):
     @classmethod
     def _get_valid_dtype(cls, series_type):
         """Return the dtype that is considered valid for a series with the given logical_type"""
-        if ks and series_type == ks.Series and cls.backup_dtype:
+        if ps and series_type == ps.Series and cls.backup_dtype:
             return cls.backup_dtype
         else:
             return cls.primary_dtype
@@ -70,8 +71,15 @@ class LogicalType(object, metaclass=LogicalTypeMetaClass):
                 raise TypeConversionError(series, new_dtype, type(self))
         return series
 
-    def validate(self, series, return_invalid_values=True):
-        return
+    def validate(self, series, *args, **kwargs):
+        """Validates that a logical type is consistent with the series dtype. Performs additional type
+        specific validation, as required. When the series' dtype does not match the logical types' required dtype,
+        raises a TypeValidationError."""
+        valid_dtype = self._get_valid_dtype(type(series))
+        if valid_dtype != str(series.dtype):
+            raise TypeValidationError(
+                f"Series dtype '{series.dtype}' is incompatible with {self.type_string} dtype."
+            )
 
 
 class Address(LogicalType):
@@ -256,7 +264,9 @@ class Datetime(LogicalType):
     def transform(self, series):
         """Converts the series data to a formatted datetime. Datetime format will be inferred if datetime_format is None."""
         new_dtype = self._get_valid_dtype(type(series))
-        if new_dtype != str(series.dtype):
+        series_dtype = str(series.dtype)
+
+        if new_dtype != series_dtype:
             self.datetime_format = self.datetime_format or _infer_datetime_format(
                 series
             )
@@ -266,9 +276,9 @@ class Datetime(LogicalType):
                     series, format=self.datetime_format, errors="coerce"
                 )
                 series.name = name
-            elif _is_koalas_series(series):
-                series = ks.Series(
-                    ks.to_datetime(
+            elif _is_spark_series(series):
+                series = ps.Series(
+                    ps.to_datetime(
                         series.to_numpy(), format=self.datetime_format, errors="coerce"
                     ),
                     name=series.name,
@@ -414,7 +424,7 @@ class LatLong(LogicalType):
 
     Note:
         LatLong values will be stored with the object dtype as a
-        tuple of floats (or a list of floats for Koalas DataFrames)
+        tuple of floats (or a list of floats for Spark DataFrames)
         and must contain only two values.
 
         Null latitude or longitude values will be stored as np.nan, and
@@ -432,21 +442,31 @@ class LatLong(LogicalType):
     primary_dtype = "object"
 
     def transform(self, series):
-        """Formats a series to be a tuple (or list for Koalas) of two floats."""
+        """Formats a series to be a tuple (or list for Spark) of two floats."""
         if _is_dask_series(series):
             name = series.name
             meta = (series, tuple([float, float]))
             series = series.apply(_reformat_to_latlong, meta=meta)
             series.name = name
-        elif _is_koalas_series(series):
+        elif _is_spark_series(series):
             formatted_series = series.to_pandas().apply(
-                _reformat_to_latlong, is_koalas=True
+                _reformat_to_latlong, is_spark=True
             )
-            series = ks.from_pandas(formatted_series)
+            series = ps.from_pandas(formatted_series)
         else:
             series = series.apply(_reformat_to_latlong)
 
         return super().transform(series)
+
+    def validate(self, series, return_invalid_values=False):
+        # TODO: we'll want to actually handle return_invalid_values in the ordinal and latlong logical types.
+        super().validate(series)
+        if not _is_valid_latlong_series(series):
+            raise TypeValidationError(
+                "Cannot initialize Woodwork. Series does not contain properly formatted "
+                "LatLong data. Try reformatting before initializing or use the "
+                "woodwork.init_series function to initialize."
+            )
 
 
 class NaturalLanguage(LogicalType):
@@ -502,7 +522,7 @@ class Ordinal(LogicalType):
     def __init__(self, order=None):
         self.order = order
 
-    def _validate_data(self, series):
+    def _validate_order_values(self, series):
         """Make sure order values are properly defined and confirm the supplied series
         does not contain any values that are not in the specified order values"""
         if self.order is None:
@@ -523,13 +543,18 @@ class Ordinal(LogicalType):
 
     def transform(self, series):
         """Validates the series and converts the dtype to match the logical type's if it is different."""
-        self._validate_data(series)
+        self._validate_order_values(series)
 
         typed_ser = super().transform(series)
         if pdtypes.is_categorical_dtype(typed_ser.dtype):
             typed_ser = typed_ser.cat.set_categories(self.order, ordered=True)
 
         return typed_ser
+
+    def validate(self, series, return_invalid_values=False):
+        # TODO: we'll want to actually handle return_invalid_values in the ordinal and latlong logical types.
+        super().validate(series)
+        self._validate_order_values(series)
 
 
 class PhoneNumber(LogicalType):
@@ -646,13 +671,13 @@ def _regex_validate(regex_key, series, return_invalid_values):
     """Validates data values based on the logical type regex in the config."""
     regex = config.get_option(regex_key)
 
-    if _is_koalas_series(series):
+    if _is_spark_series(series):
 
         def match(x):
             if isinstance(x, str):
                 return bool(re.match(regex, x))
 
-        invalid = ~series.apply(match).astype("boolean")
+        invalid = series.apply(match).astype("boolean") == False  # noqa: E712
 
     else:
         invalid = ~series.str.match(regex).astype("boolean")