+++ woodwork/tests/logical_types/test_logical_types.py
@@ -3,7 +3,7 @@ import re
 import pandas as pd
 import pytest
 
-from woodwork.accessor_utils import _is_koalas_series
+from woodwork.accessor_utils import _is_spark_series, init_series
 from woodwork.exceptions import TypeConversionWarning, TypeValidationError
 from woodwork.logical_types import (
     URL,
@@ -20,7 +20,7 @@ from woodwork.logical_types import (
 from woodwork.tests.testing_utils.table_utils import to_pandas
 from woodwork.utils import import_or_none
 
-ks = import_or_none("databricks.koalas")
+ps = import_or_none("pyspark.pandas")
 
 
 def test_logical_eq():
@@ -90,17 +90,17 @@ def test_ordinal_transform_dask(ordinal_transform_series_dask) -> None:
     pd.testing.assert_index_equal(ser_.cat.categories, pd.Int64Index(order))
 
 
-def test_ordinal_transform_koalas(ordinal_transform_series_koalas) -> None:
+def test_ordinal_transform_spark(ordinal_transform_series_spark) -> None:
     order = [2, 1, 3]
     typ = Ordinal(order=order)
-    ser_ = typ.transform(ordinal_transform_series_koalas)
+    ser_ = typ.transform(ordinal_transform_series_spark)
 
     assert ser_.dtype == pd.StringDtype()
 
 
 def test_get_valid_dtype(sample_series):
     valid_dtype = Categorical._get_valid_dtype(type(sample_series))
-    if _is_koalas_series(sample_series):
+    if _is_spark_series(sample_series):
         assert valid_dtype == "string"
     else:
         assert valid_dtype == "category"
@@ -112,7 +112,7 @@ def test_get_valid_dtype(sample_series):
 def test_latlong_transform(latlong_df):
     df_type = str(type(latlong_df))
     dask = "dask" in df_type
-    koalas = "koalas" in df_type
+    spark = "spark" in df_type
     nan = float("nan")
 
     expected_data = {
@@ -134,11 +134,25 @@ def test_latlong_transform(latlong_df):
 
         if dask:
             actual = actual.compute()
-        elif koalas:
+        elif spark:
             actual = actual.to_pandas()
 
         expected = pd.Series(expected_data[column], name=column)
         pd.testing.assert_series_equal(actual, expected)
+
+
+def test_latlong_validate(latlong_df):
+    error_message = re.escape(
+        "Cannot initialize Woodwork. Series does not contain properly formatted "
+        "LatLong data. Try reformatting before initializing or use the "
+        "woodwork.init_series function to initialize."
+    )
+    latlong = LatLong()
+    series = latlong_df["tuple_ints"]
+    new_series = init_series(series, logical_type=LatLong)
+    latlong.validate(new_series)
+    with pytest.raises(TypeValidationError, match=error_message):
+        latlong.validate(series)
 
 
 def test_datetime_transform(datetimes):
@@ -185,11 +199,11 @@ def test_datetime_coerce_user_format():
 def test_ordinal_transform(sample_series):
     series_type = str(type(sample_series))
     dask = "dask" in series_type
-    koalas = "koalas" in series_type
+    spark = "spark" in series_type
 
-    if dask or koalas:
+    if dask or spark:
         pytest.xfail(
-            "Fails with Dask and Koalas - ordinal data validation not supported"
+            "Fails with Dask and Spark - ordinal data validation not supported"
         )
 
     ordinal_incomplete_order = Ordinal(order=["a", "b"])
@@ -202,17 +216,45 @@ def test_ordinal_transform(sample_series):
         ordinal_incomplete_order.transform(sample_series)
 
 
+def test_ordinal_validate(sample_series):
+    series_type = str(type(sample_series))
+    dask = "dask" in series_type
+    spark = "spark" in series_type
+
+    if dask or spark:
+        pytest.xfail(
+            "Fails with Dask and Spark - ordinal data validation not supported"
+        )
+
+    ordinal_incomplete_order = Ordinal(order=["a", "b"])
+    error_msg = re.escape(
+        "Ordinal column sample_series contains values that are not "
+        "present in the order values provided: ['c']"
+    )
+
+    with pytest.raises(ValueError, match=error_msg):
+        ordinal_incomplete_order.validate(sample_series)
+
+    new_type = "string"
+    error_message = re.escape(
+        f"Series dtype '{new_type}' is incompatible with ordinal dtype."
+    )
+    with pytest.raises(TypeValidationError, match=error_message):
+        ordinal_incomplete_order.validate(sample_series.astype(new_type))
+
+
 def test_email_address_validate(sample_df):
-    series = sample_df["email"]
-    invalid_row = pd.Series({4: "bad_email"}, name="email")
-
-    if _is_koalas_series(series):
-        invalid_row = ks.from_pandas(invalid_row)
-
     email_address = EmailAddress()
+    dtype = email_address.primary_dtype
+    series = sample_df["email"].astype(dtype)
+    invalid_row = pd.Series({4: "bad_email"}, name="email").astype(dtype)
+
+    if _is_spark_series(series):
+        invalid_row = ps.from_pandas(invalid_row)
+
     assert email_address.validate(series) is None
 
-    series = series.append(invalid_row)
+    series = series.append(invalid_row).astype(dtype)
     match = "Series email contains invalid email address values. "
     match += "The email_inference_regex can be changed in the config if needed."
 
@@ -220,21 +262,21 @@ def test_email_address_validate(sample_df):
         email_address.validate(series)
 
     actual = email_address.validate(series, return_invalid_values=True)
-    expected = pd.Series({4: "bad_email"}, name="email")
+    expected = pd.Series({4: "bad_email"}, name="email").astype(dtype)
     assert to_pandas(actual).equals(expected)
 
 
 def test_url_validate(sample_df):
-    series = sample_df["url"]
-    invalid_row = pd.Series({4: "bad_url"}, name="url")
-
-    if _is_koalas_series(series):
-        invalid_row = ks.from_pandas(invalid_row)
-
     logical_type = URL()
+    dtype = logical_type.primary_dtype
+    series = sample_df["url"].astype(dtype)
+    invalid_row = pd.Series({4: "bad_url"}, name="url").astype(dtype)
+    if _is_spark_series(series):
+        invalid_row = ps.from_pandas(invalid_row)
+
     assert logical_type.validate(series) is None
 
-    series = series.append(invalid_row)
+    series = series.append(invalid_row).astype(dtype)
     match = "Series url contains invalid url values. "
     match += "The url_inference_regex can be changed in the config if needed."
 
@@ -242,7 +284,7 @@ def test_url_validate(sample_df):
         logical_type.validate(series)
 
     actual = logical_type.validate(series, return_invalid_values=True)
-    expected = pd.Series({4: "bad_url"}, name="url")
+    expected = pd.Series({4: "bad_url"}, name="url").astype(dtype)
     assert to_pandas(actual).equals(expected)
 
 
@@ -262,8 +304,8 @@ def test_age_validate(sample_df, logical_type):
     assert logical_type.validate(series, return_invalid_values=False) is None
     invalid_row = pd.Series({4: -3}, name="age", dtype=dtype)
 
-    if _is_koalas_series(series):
-        invalid_row = ks.from_pandas(invalid_row)
+    if _is_spark_series(series):
+        invalid_row = ps.from_pandas(invalid_row)
     series = series.append(invalid_row).astype(dtype)
 
     match = "Series age contains negative values."